fix- refactor proxy place internal user endpoints in sep file
