+++ litellm/tests/test_blocked_user_list.py
@@ -28,10 +28,7 @@ from litellm.proxy.utils import PrismaClient, ProxyLogging, hash_token
 import pytest, logging, asyncio
 import litellm, asyncio
 from litellm.proxy.proxy_server import (
-    new_user,
     user_api_key_auth,
-    user_update,
-    user_info,
     block_user,
 )
 from litellm.proxy.management_endpoints.key_management_endpoints import (
@@ -40,6 +37,11 @@ from litellm.proxy.management_endpoints.key_management_endpoints import (
     update_key_fn,
     generate_key_fn,
     generate_key_helper_fn,
+)
+from litellm.proxy.management_endpoints.internal_user_endpoints import (
+    new_user,
+    user_update,
+    user_info,
 )
 from litellm.proxy.spend_reporting_endpoints.spend_management_endpoints import (
     spend_user_fn,
