+++ litellm/tests/test_key_generate_prisma.py
@@ -37,10 +37,7 @@ sys.path.insert(
 import pytest, logging, asyncio
 import litellm, asyncio
 from litellm.proxy.proxy_server import (
-    new_user,
     user_api_key_auth,
-    user_update,
-    user_info,
     chat_completion,
     completion,
     embeddings,
@@ -56,6 +53,11 @@ from litellm.proxy.management_endpoints.key_management_endpoints import (
     update_key_fn,
     generate_key_fn,
     generate_key_helper_fn,
+)
+from litellm.proxy.management_endpoints.internal_user_endpoints import (
+    new_user,
+    user_update,
+    user_info,
 )
 from litellm.proxy.management_endpoints.team_endpoints import (
     team_info,
